% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlpanel.R
\name{dmlpanel}
\alias{dmlpanel}
\title{DML for High-Dimensional Linear Panel Data Models in R}
\usage{
dmlpanel(
  y,
  v,
  w,
  id,
  C1 = NULL,
  C2 = NULL,
  Omega = NULL,
  S2 = NULL,
  TotT,
  L = 5,
  re = 1.1,
  indreg
)
}
\arguments{
\item{y}{An nT x 1 vector of outcome variables. It must be provided.}

\item{v}{An nT x q matrix of regressors with "random" coefficients. It must be provided.}

\item{w}{An nT x p matrix of regressors with common coefficients. It must be provided.}

\item{id}{An nT x 1 vector of repeated id's. It must be provided.}

\item{C1}{A p x p1 matrix. Defult is NULL.}

\item{C2}{A q x p2 matrix.D efult is NULL.}

\item{Omega}{A q x q matrix.}

\item{S2}{A selection matrix for modeling vec(Sigma) (that is, the var-cov matrix of idiosyncratic errors). The number of columns of this matrix determines the dimension of omega_0 and thus the degree of parametrization of the variance of errors. If this matrix is not provided, iid errors are assumed (default).}

\item{TotT}{Number of "periods" in the panel (i.e., number of times a unit in the cross-section dimension is observed). It must be provided.}

\item{L}{number of folds (scalar). Default is 5.}

\item{re}{regularization constant (scalar). Default is 1.1.}

\item{indreg}{A natural number indicating inference will be conducted for the first indreg regressors in W.It must be provided.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list with parameter estimates and standard errors.
}\if{html}{\out{</div>}}
}
\description{
\code{dmlpanel()} implements Debiased Machine Learning (DML) for high-dimensional linear panel data models with common parameters and random coefficients, as introduced by "Debiased Machine Learning Inference for Functionals of Unoberseved Heterogeneity", by Argañaraz, F., and Escanciano, J.C., Argañaraz and Escanciano (2025).
}
